# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Complete Ecormmerce server.",
        "title": "Gin Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/user/addtocart": {
            "post": {
                "description": "add products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "Add product into user cart",
                "operationId": "add-to-cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email of the user",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "description": "User details",
                        "name": "user_details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Productdetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/checkout": {
            "get": {
                "description": "checkout",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "Checkout",
                "operationId": "checkout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the user",
                        "name": "user",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "wallet of the user",
                        "name": "wallet",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "address id of the user",
                        "name": "addressID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "payment method  of the user",
                        "name": "PaymentMethod",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/checkoutAddress": {
            "post": {
                "description": "address for checkout",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "Checkout address",
                "operationId": "checkout address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email of the user",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the user",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "phone number of the user",
                        "name": "phone_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pincode of the user",
                        "name": "pincode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "area of the user",
                        "name": "area",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "house of the user",
                        "name": "house",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "landmark of the user",
                        "name": "landmark",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "city of the user",
                        "name": "city",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/viewcart": {
            "get": {
                "description": "add products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "view cart",
                "operationId": "view-cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email of the user",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "product of the user",
                        "name": "productid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "product of the user",
                        "name": "quantity",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Productdetails": {
            "type": "object",
            "properties": {
                "productID": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
    }
}